+++ src/js/visual/configurator-csr.js
@@ -91,6 +91,7 @@ ripe.ConfiguratorCsr.prototype.init = function() {
     this.animations = [];
     this.isChangeFrameAnimationRunning = false;
     this._pendingOps = [];
+    this._postRenderCallback = null;
 
     // CSR variables
     this.rendererOptions = null;
@@ -111,7 +112,10 @@ ripe.ConfiguratorCsr.prototype.init = function() {
         renderedInitials: null,
         mesh: null,
         baseTexture: null,
-        displacementTexture: null
+        displacementTexture: null,
+        metallicTexture: null,
+        normalTexture: null,
+        roughnessTexture: null
     };
 
     // CSR debug variables
@@ -526,6 +530,22 @@ ripe.ConfiguratorCsr.prototype.prcFrame = async function() {
 };
 
 /**
+ * Sets the callback function for the post render call.
+ *
+ * @param {Function} callback The function to be called.
+ */
+ripe.ConfiguratorCsr.prototype.setPostRender = function(callback) {
+    this._postRenderCallback = callback;
+};
+
+/**
+ * Clears the post render callback registry.
+ */
+ripe.ConfiguratorCsr.prototype.unsetPostRender = function() {
+    this._postRenderCallback = null;
+};
+
+/**
  * Tries to obtain the best possible size for the configurator
  * defaulting to the client with of the element as fallback.
  *
@@ -805,15 +825,38 @@ ripe.ConfiguratorCsr.prototype._initCsrRenderedInitials = function() {
 
     // apply textures to initials
     if (this.initialsRefs.baseTexture) {
-        this.initialsRefs.renderedInitials.setBaseTexture(
+        this.initialsRefs.renderedInitials.setTexture(
+            "base",
             this.initialsRefs.baseTexture,
             this.initialsOptions.options.baseTextureOptions
         );
     }
     if (this.initialsRefs.displacementTexture) {
-        this.initialsRefs.renderedInitials.setDisplacementTexture(
+        this.initialsRefs.renderedInitials.setTexture(
+            "displacement",
             this.initialsRefs.displacementTexture,
             this.initialsOptions.options.displacementTextureOptions
+        );
+    }
+    if (this.initialsRefs.metallicTexture) {
+        this.initialsRefs.renderedInitials.setTexture(
+            "metallic",
+            this.initialsRefs.metallicTexture,
+            this.initialsOptions.options.metallicTextureOptions
+        );
+    }
+    if (this.initialsRefs.normalTexture) {
+        this.initialsRefs.renderedInitials.setTexture(
+            "normal",
+            this.initialsRefs.normalTexture,
+            this.initialsOptions.options.normalTextureOptions
+        );
+    }
+    if (this.initialsRefs.roughnessTexture) {
+        this.initialsRefs.renderedInitials.setTexture(
+            "roughness",
+            this.initialsRefs.roughnessTexture,
+            this.initialsOptions.options.roughnessTextureOptions
         );
     }
 
@@ -863,6 +906,9 @@ ripe.ConfiguratorCsr.prototype._unloadCsrRenderedInitialsResources = function() 
     // cleanup loaded textures
     if (this.initialsRefs.baseTexture) this.initialsRefs.baseTexture.dispose();
     if (this.initialsRefs.displacementTexture) this.initialsRefs.displacementTexture.dispose();
+    if (this.initialsRefs.metallicTexture) this.initialsRefs.metallicTexture.dispose();
+    if (this.initialsRefs.normalTexture) this.initialsRefs.normalTexture.dispose();
+    if (this.initialsRefs.roughnessTexture) this.initialsRefs.roughnessTexture.dispose();
 
     // free all resources used
     if (this.initialsRefs.renderedInitials) this.initialsRefs.renderedInitials.destroy();
@@ -870,7 +916,10 @@ ripe.ConfiguratorCsr.prototype._unloadCsrRenderedInitialsResources = function() 
         renderedInitials: null,
         mesh: null,
         baseTexture: null,
-        displacementTexture: null
+        displacementTexture: null,
+        metallicTexture: null,
+        normalTexture: null,
+        roughnessTexture: null
     };
 };
 
@@ -1168,6 +1217,7 @@ ripe.ConfiguratorCsr.prototype._render = function() {
     if (!this.camera) throw new Error("Camera not initiated");
     if (!this.renderer) throw new Error("Renderer not initiated");
     this.renderer.render(this.scene, this.camera);
+    this._onPostRender();
 };
 
 /**
@@ -1239,6 +1289,13 @@ ripe.ConfiguratorCsr.prototype._onAnimationLoop = function(self) {
 
     // renders a frame
     self._render();
+};
+
+/**
+ * @ignore
+ */
+ripe.ConfiguratorCsr.prototype._onPostRender = function() {
+    if (this._postRenderCallback) this._postRenderCallback();
 };
 
 /**
@@ -1364,9 +1421,15 @@ ripe.ConfiguratorCsr.prototype._onPostConfig = async function(self, config) {
         // checks if it should load assets used by the initials
         let baseTexturePath = null;
         let displacementTexturePath = null;
+        let metallicTexturePath = null;
+        let normalTexturePath = null;
+        let roughnessTexturePath = null;
         if (initialsEnabled) {
-            baseTexturePath = this.owner.getInitials3dBaseTextureUrl();
-            displacementTexturePath = this.owner.getInitials3dDisplacementTextureUrl();
+            baseTexturePath = this.owner.getTextureMapUrl("pattern");
+            displacementTexturePath = this.owner.getTextureMapUrl("displacement");
+            metallicTexturePath = this.owner.getTextureMapUrl("metallic");
+            normalTexturePath = this.owner.getTextureMapUrl("normal");
+            roughnessTexturePath = this.owner.getTextureMapUrl("roughness");
         }
 
         // loads assets
@@ -1374,12 +1437,18 @@ ripe.ConfiguratorCsr.prototype._onPostConfig = async function(self, config) {
             this.mesh,
             this.environmentTexture,
             this.initialsRefs.baseTexture,
-            this.initialsRefs.displacementTexture
+            this.initialsRefs.displacementTexture,
+            this.initialsRefs.metallicTexture,
+            this.initialsRefs.normalTexture,
+            this.initialsRefs.roughnessTexture
         ] = await Promise.all([
             this._loadMesh(meshPath, meshFormat),
             envPath ? ripe.CsrUtils.loadEnvironment(envPath, envFormat) : null,
             baseTexturePath ? ripe.CsrUtils.loadTexture(baseTexturePath) : null,
-            displacementTexturePath ? ripe.CsrUtils.loadTexture(displacementTexturePath) : null
+            displacementTexturePath ? ripe.CsrUtils.loadTexture(displacementTexturePath) : null,
+            metallicTexturePath ? ripe.CsrUtils.loadTexture(metallicTexturePath) : null,
+            normalTexturePath ? ripe.CsrUtils.loadTexture(normalTexturePath) : null,
+            roughnessTexturePath ? ripe.CsrUtils.loadTexture(roughnessTexturePath) : null
         ]);
 
         // gets the 3d set from the config